requirements.txt
streamlit==1.26.0
openai
streamlit_custom_notification_box
PyPDF2 
python-docx
